Created custom `SortIndex` type, based on new optimizable integer type. Added some compiler infrastructure, including most of the compiler for the free theory. Added some automaton implementation.

+++ mod2-lib/src/api/variable_theory/mod.rs
@@ -8,7 +8,7 @@ pub use variable_symbol::*;
 pub use variable_dag_node::*;
 
 /// The type used to store variable indices.
-pub type VariableIndex = i32;
+pub type VariableIndex = usize;
 
 /// The `VariableType` of a variable determines what the variable is able to bind to. A `Blank` variable binds to a
 /// single `Term`, a `Sequence` variable binds to a sequence of one or more `Term`s, and a `NullSequence` binds to a
